+++ src/opticalc/integrated.py
@@ -1,113 +1,28 @@
-from functools import partial
-
 import logging
-import pywincalc
-from py_igsdb_optical_data.standard import CalculationStandardMethodTypes
 
-from opticalc.product import Product, ProductSubtype
-from py_igsdb_optical_data.optical import OpticalStandardMethodResults, OpticalColorResults, \
-    IntegratedSpectralAveragesSummaryValues, OpticalColorFluxResults, OpticalColorResult, RGBResult, \
-    LabResult, TrichromaticResult, ThermalIRResults, OpticalStandardMethodFluxResults, \
+import pywincalc
+from py_igsdb_base_data.optical import OpticalStandardMethodResults, OpticalColorResults, \
+    IntegratedSpectralAveragesSummaryValues, OpticalColorFluxResults, OpticalColorResult, ThermalIRResults, \
+    OpticalStandardMethodFluxResults, \
     IntegratedSpectralAveragesSummaryValuesFactory
+from py_igsdb_base_data.standard import CalculationStandardMethodTypes
+
+from opticalc.exceptions import SpectralAveragesSummaryCalculationException
+from opticalc.product import Product
+from opticalc.util import convert_trichromatic_result, convert_lab_result, convert_rgb_result, convert_product
 
 logger = logging.getLogger(__name__)
 
 
-def convert_wavelength_data(raw_wavelength_data):
-    pywincalc_wavelength_measured_data = []
-    for individual_wavelength_measurement in raw_wavelength_data:
-        wavelength = individual_wavelength_measurement["w"]
-        # In this case the raw data only has the direct component measured
-        # Diffuse measured data is also not yet supported in the calculations
-        direct_component = pywincalc.OpticalMeasurementComponent(
-            individual_wavelength_measurement["specular"]["tf"],
-            individual_wavelength_measurement["specular"]["tb"],
-            individual_wavelength_measurement["specular"]["rf"],
-            individual_wavelength_measurement["specular"]["rb"])
-        pywincalc_wavelength_measured_data.append(pywincalc.WavelengthData(wavelength, direct_component))
-
-    return pywincalc_wavelength_measured_data
-
-
-def convert_subtype(subtype):
-    subtype_mapping = {ProductSubtype.MONOLITHIC: pywincalc.MaterialType.MONOLITHIC,
-                       ProductSubtype.APPLIED_FILM: pywincalc.MaterialType.APPLIED_FILM,
-                       ProductSubtype.COATED: pywincalc.MaterialType.COATED,
-                       ProductSubtype.LAMINATE: pywincalc.MaterialType.LAMINATE,
-                       ProductSubtype.INTERLAYER: pywincalc.MaterialType.INTERLAYER,
-                       ProductSubtype.FILM: pywincalc.MaterialType.FILM}
-
-    pywincalc_material = subtype_mapping.get(subtype)
-    if pywincalc_material is None:
-        raise RuntimeError("Unsupported subtype: {t}".format(t=subtype))
-    return pywincalc_material
-
-
-def convert_coated_side(coated_side: str) -> pywincalc.CoatedSide:
-    if not coated_side:
-        return pywincalc.CoatedSide.NEITHER
-    coated_side = coated_side.upper()
-    mapping = {
-        "FRONT": pywincalc.CoatedSide.FRONT,
-        "BACK": pywincalc.CoatedSide.BACK,
-        "BOTH": pywincalc.CoatedSide.BOTH,
-        "NEITHER": pywincalc.CoatedSide.NEITHER,
-        "NA": pywincalc.CoatedSide.NEITHER
-    }
+def calc_optical(glazing_system: pywincalc.GlazingSystem, method_name: str) -> OpticalStandardMethodResults:
     try:
-        return mapping.get(coated_side)
+        CalculationStandardMethodTypes[method_name]
     except KeyError:
-        raise RuntimeError(f"Unsupported coated side: {coated_side}")
+        raise ValueError(f"Invalid method: {method_name}")
 
-
-def convert_product(product) -> pywincalc.ProductDataOpticalAndThermal:
-    """
-    Converts a product.Product dataclass instance into a
-    dataclass instance that PyWinCalc can work with.
-
-    :param product:     Instance of a populated product.Product dataclass
-
-    :return:
-    Instance of pywincalc.ProductDataOpticalAndThermal
-    """
-
-    optical_data = product.physical_properties.optical_properties.optical_data
-    if not optical_data:
-        raise Exception(f"No optical_data is defined on product: {product}")
-
-    try:
-        wavelength_data = optical_data["angle_blocks"][0]["wavelength_data"]
-        if not wavelength_data:
-            raise Exception("No wavelength data")
-    except Exception as e:
-        raise Exception("Could not find wavelength data in product : {product}") from e
-
-    wavelength_data = convert_wavelength_data(wavelength_data)
-    material_type = convert_subtype(product.subtype)
-    material_thickness = product.physical_properties.thickness
-
-    # We use the top-level properties on 'product' to get emissivity and TIR.
-    # These property methods will return a predefined value if one exists,
-    # otherwise will return a calculated value if one exists.
-    emissivity_front = product.emissivity_front
-    emissivity_back = product.emissivity_back
-    ir_transmittance_front = product.tir_front
-    ir_transmittance_back = product.tir_back
-
-    coated_side = convert_coated_side(product.coated_side)
-
-    optical_data = pywincalc.ProductDataOpticalNBand(material_type, material_thickness, wavelength_data, coated_side,
-                                                     ir_transmittance_front, ir_transmittance_back,
-                                                     emissivity_front, emissivity_back)
-
-    layer = pywincalc.ProductDataOpticalAndThermal(optical_data, None)
-    return layer
-
-
-def calc_optical(glazing_system, method) -> OpticalStandardMethodResults:
     translated_results = OpticalStandardMethodResults()
     try:
-        results = glazing_system.optical_method_results(method)
+        results = glazing_system.optical_method_results(method_name)
         system_results = results.system_results
         translated_results.transmittance_front = OpticalStandardMethodFluxResults(
             direct_direct=system_results.front.transmittance.direct_direct,
@@ -142,26 +57,28 @@ def calc_optical(glazing_system, method) -> OpticalStandardMethodResults:
         translated_results.absorptance_front_hemispheric = results.layer_results[0].front.absorptance.diffuse
         translated_results.absorptance_back_hemispheric = results.layer_results[0].back.absorptance.diffuse
 
-
     except Exception as e:
         translated_results.error = e
 
     return translated_results
 
 
-def convert_trichromatic_result(trichromatic):
-    return TrichromaticResult(x=trichromatic.X, y=trichromatic.Y, z=trichromatic.Z)
+def calc_color(glazing_system: pywincalc.GlazingSystem) -> OpticalColorResults:
+    """
+    Uses pywincalc to generate color information for a given glazing system.
+    Returns information in a populated instance of the OpticalColorResults dataclass
 
+    Args:
+        glazing_system: Instance of a pywincalc GlazingSystem
 
-def convert_lab_result(lab):
-    return LabResult(l=lab.L, a=lab.a, b=lab.b)
+    Returns:
+        An instance of OpticalColorResults dataclass populated with results.
 
+    Raises:
+        SpectralAveragesSummaryCalculationException if an error is encountered when generating or parsing
+        results.
+    """
 
-def convert_rgb_result(rgb):
-    return RGBResult(r=rgb.R, g=rgb.G, b=rgb.B)
-
-
-def calc_color(glazing_system) -> OpticalColorResults:
     translated_results = OpticalColorResults()
     try:
         results = glazing_system.color()
@@ -332,7 +249,8 @@ def calc_color(glazing_system) -> OpticalColorResults:
     return translated_results
 
 
-def generate_thermal_ir_results(pywincalc_layer, optical_standard) -> ThermalIRResults:
+def generate_thermal_ir_results(optical_standard: pywincalc.OpticalStandard,
+                                pywincalc_layer: pywincalc.ProductDataOpticalAndThermal) -> ThermalIRResults:
     translated_results = ThermalIRResults()
     pywincalc_results = pywincalc.calc_thermal_ir(optical_standard, pywincalc_layer)
     translated_results.transmittance_front_diffuse_diffuse = pywincalc_results.transmittance_front_diffuse_diffuse
@@ -342,27 +260,32 @@ def generate_thermal_ir_results(pywincalc_layer, optical_standard) -> ThermalIRR
     return translated_results
 
 
-class SpectralAveragesSummaryCalculationException(Exception):
-    """
-    Capture info about which part of integrated spectral averages
-    calculation process went bad.
-    """
-    pass
-
-
 def generate_integrated_spectral_averages_summary(product: Product,
                                                   optical_standard: pywincalc.OpticalStandard) \
         -> IntegratedSpectralAveragesSummaryValues:
     """
-    Generate integrated spectral averages summary for a given product and standard.
+    Uses pywincalc to generate an integrated spectral averages summary for a given product
+    and standard.
+
+    Args:
+        product:            Instance of a product dataclass
+        optical_standard:   Instance of a pywincalc OpticalStandard class.
+
+    Returns:
+        An instance of IntegratedSpectralAveragesSummaryValues dataclass populated with results.
+
+    Raises:
+        SpectralAveragesSummaryCalculationException if an error is encountered when generating or parsing
+        results.
     """
 
     summary_results: IntegratedSpectralAveragesSummaryValues = IntegratedSpectralAveragesSummaryValuesFactory.create()
-    pywincalc_layer = convert_product(product)
-    glazing_system = pywincalc.GlazingSystem(optical_standard=optical_standard, solid_layers=[pywincalc_layer])
+    pywincalc_layer: pywincalc.ProductDataOpticalAndThermal = convert_product(product)
+    glazing_system: pywincalc.GlazingSystem = pywincalc.GlazingSystem(optical_standard=optical_standard,
+                                                                      solid_layers=[pywincalc_layer])
 
     for method_name in [item.name for item in CalculationStandardMethodTypes]:
-        # Only calculate results for a method if its supported in the current optical standard...
+        # Only calculate results for a method if it's supported in the current optical standard...
         if method_name in optical_standard.methods:
             try:
                 results: OpticalStandardMethodResults = calc_optical(glazing_system, method_name)
@@ -386,7 +309,7 @@ def generate_integrated_spectral_averages_summary(product: Product,
         raise SpectralAveragesSummaryCalculationException(error_msg) from e
 
     try:
-        summary_results.thermal_ir = generate_thermal_ir_results(pywincalc_layer, optical_standard)
+        summary_results.thermal_ir = generate_thermal_ir_results(optical_standard, pywincalc_layer)
     except Exception as e:
         error_msg = f"generate_thermal_ir_results() call failed for " \
                     f"layer: {pywincalc_layer} " \