+++ tests/test_integrated_spectral_averages_summary_calculation.py
@@ -3,7 +3,7 @@ import os
 from math import isclose
 
 import pywincalc
-from py_igsdb_optical_data.optical import IntegratedSpectralAveragesSummaryValues
+from py_igsdb_base_data.optical import IntegratedSpectralAveragesSummaryValues
 
 from opticalc.integrated import generate_integrated_spectral_averages_summary
 from opticalc.product import Product
@@ -62,8 +62,3 @@ def test_calculate_emissivity_from_wavelengths():
     assert values.thermal_ir.transmittance_front_diffuse_diffuse == 0
     assert values.thermal_ir.transmittance_back_diffuse_diffuse == 0
     
-
-if __name__ == "__main__":
-    test_generate_integrated_spectral_averages_summary()
-    test_calculate_emissivity_from_wavelengths()
-