+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 Release Notes:
 
 0.0.5
-- updated pywincalc to 2.4.2 
-  - This pywincal version has improved calculation of integrated spectral averages for shading layers with BSDF
+- Updated pywincalc to 2.4.2 
+  - This pywincalc version has improved calculation of integrated spectral averages for shading layers with BSDF
+
+0.0.6
+- Updates integrated summary values models to reflect changes in how Checkertool structures these values. 
+- Updates ProductSubtype dataclass to match Checkertool.
+- Update Product dataclass to use strings rather than enums, but validate those string values against values defined in enums.
+- Add tests.
+
+0.0.7
+- Update pywincalc dependency